==> 01-preparation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# https://www.kaggle.com/timmate/avocado-prices-2020

df = pd.read_csv('./data_csv/avocado-updated-2020.csv', parse_dates=['date'])

df.head()

2819.5+28287.42+49.9

df.shape

df['year'].unique()

jumlah = []

for i in df.columns:
    jumlah.append(df[i].nunique())

dfjumlah = pd.DataFrame(jumlah)
dfcol = pd.DataFrame(df.columns)  

unique = pd.concat([dfcol, dfjumlah], axis=1)
unique.columns = ['features', 'unique values']

unique

df[df['year'] == 2020].max()

df.isnull().sum()

df.shape

df['type'].value_counts()

df['average_price'].describe()

df['year'].value_counts().sort_index()

df.head()

df.describe().T

df.describe(include='O').T

desc = []

for i in df.columns:
    desc.append([
        i,
        df[i].dtypes,
        df[i].isnull().sum(),
        round((((df[i].isnull().sum())/(len(df)))*100), 2),
        df[i].nunique(),
        df[i].drop_duplicates().sample(2).values
    ])

dfDesc = pd.DataFrame(desc, columns=['dataFeatures', 'dataType', 'null', 'nullPercentage', 'unique', 'uniqueSample'])

dfDesc

## insight data:

ada 13 kolom dalam dataset:
1. tanggal: dari januari 2015 - mei 2020
2. average_price: harga satuan avocado (US dollar)
3. total_volume: total penjualan avocado (lbs)
4. 4046
5. 4225
6. 4770 ==> total volume penjualan avocado PLU 4470 (perbedaan antara PLU 4046, 4225, 4770 adalah ukuran avocado)
7. total_bags: total kantong avocado terjual
8. small_bags: total penjualan avocado dengan pembelian menggunakan kantong kecil (tidak diketahui ukuran/dimensi kantong kecil)
9. large_bags: total penjualan avocado dengan pembelian menggunakan kantong large (tidak diketahui ukuran/dimensi kantong besar)
10. xlarge_bags: total penjualan avocado dengan pembelian menggunakan kantong extra large (tidak diketahui ukuran/dimensi kantong extra large)
11. type: tipe avocado (conventional dan organik)
12. year: tahun pencatatan transaksi: 2015 - 2020
13. geography: region kota tempat transaksi

- tidak ada data null dalam semua kolom
- dataset terdiri dari 13 kolom dan 30021 baris

Data dapat dilengkapi dengan keterangan Region di USA: https://www.nationalgeographic.org/maps/united-states-regions/#:~:text=A%20common%20way%20of%20referring,West%2C%20Southeast%2C%20and%20Midwest

## Problem statement:
Suatu online grocery shopping platform ingin mengekspansi wilayah layanannya (servis area) di kota di U.S. Berikan gambaran bagaimana kondisi pasar avocado di kota-kota U.S. Dapatkah algoritma machine learning membantu decision making perusahaan terhadap pertimbangan ekspansi ke kota baru?

## Tujuan:
- Analisis price elasticity dan volatility avocado di region-region, ataupun kota-kota di U.S
- Tentukan apakah pasar avocado di U.S dipengaruhi faktor musiman
- Buat machine learning model untuk memprediksi harga avocado guna membantu proses decision making perusahaan

## Data Preparation

terdapat beberapa kemungkinan kondisi pada dateset yang raw (mentah):
- adanya missing value
- adanya data yang tidak konsisten
- terdapat outliers
- data yang irrelevan terhadap target

sehingga pada tahap awal perlu dilakukan data cleaning, khususnya pengecekan terhadap adanya missing value dan data yang tidak konsisten.

missing value susah untuk diinterpretasikan oleh machine

### ** ekstrak data pada kolom 'date' **

# df['month'] = df['date'].dt.month

df['Date'] = df['date'].dt.day

df['month'] = df['date'].dt.month_name()

df.head()

df['month_num'] = df['date'].dt.month

df.drop(columns=['date'], inplace=True)

df.head()

### ** Tambah keterangan kolom region pada df **

print(df['geography'].unique())

df['geography'].nunique()

terdapat data yang inkonsisten dalam kolom 'geography' >> selain data nama-nama kota di AS, di kolom ini juga terdapat data 'Total_US', dan region-region di AS (Midsouth, Northeast, Southeast, South Central, West) >> didrop sehingga mendapatkan data keseluruhan hanya city

df_clean = df[(df['geography'] != 'Total U.S.') & (df['geography'] != 'Midsouth') & (df['geography'] != 'Northeast') & (df['geography'] != 'South Central') & (df['geography'] != 'Southeast')
& (df['geography'] != 'West')]

df_clean

df_clean['geography'].unique()

def region(df_clean):    
    if df_clean['geography'] == 'Albany' or df_clean['geography'] == 'Baltimore/Washington' or df_clean['geography'] == 'Boston' or df_clean['geography'] == 'Buffalo/Rochester' or df_clean['geography'] == 'Harrisburg/Scranton' or df_clean['geography'] == 'Hartford/Springfield' or df_clean['geography'] == 'New York' or df_clean['geography'] == 'Northern New England' or df_clean['geography'] == 'Philadelphia' or df_clean['geography'] == 'Pittsburgh' or df_clean['geography'] == 'Syracuse':
        return 'Northeastern'
    if df_clean['geography'] == 'Atlanta' or df_clean['geography'] == 'Charlotte' or df_clean['geography'] == 'Jacksonville' or df_clean['geography'] == 'Louisville' or df_clean['geography'] == 'Miami/Ft. Lauderdale' or df_clean['geography'] == 'Nashville' or df_clean['geography'] == 'New Orleans/Mobile' or df_clean['geography'] == 'Orlando' or df_clean['geography'] == 'Raleigh/Greensboro' or df_clean['geography'] == 'Richmond/Norfolk' or df_clean['geography'] == 'Roanoke' or df_clean['geography'] == 'South Carolina' or df_clean['geography'] == 'Tampa':
        return 'Southeastern'
    if df_clean['geography'] == 'Boise' or df_clean['geography'] == 'California' or df_clean['geography'] == 'Denver' or df_clean['geography'] == 'Las Vegas' or df_clean['geography'] == 'Los Angeles' or df_clean['geography'] == 'Plains' or df_clean['geography'] == 'Portland' or df_clean['geography'] == 'Sacramento' or df_clean['geography'] == 'San Diego' or df_clean['geography'] == 'San Francisco' or df_clean['geography'] == 'Seattle' or df_clean['geography'] == 'Spokane':
        return 'Western'
    if df_clean['geography'] == 'Chicago' or df_clean['geography'] == 'Cincinnati/Dayton' or df_clean['geography'] == 'Columbus' or df_clean['geography'] == 'Detroit' or df_clean['geography'] == 'Grand Rapids' or df_clean['geography'] == 'Great Lakes' or df_clean['geography'] == 'Indianapolis' or df_clean['geography'] == 'St. Louis':
        return 'Midwestern'
    if df_clean['geography'] == 'Dallas/Ft. Worth' or df_clean['geography'] == 'Houston' or df_clean['geography'] == 'Phoenix/Tucson' or df_clean['geography'] == 'West Tex/New Mexico':
        return 'Southwestern'
#     return 'Other'

# df_clean['region'] = df_clean.apply(lambda df_clean: region(df_clean), axis=1)

region = df_clean.apply(lambda df_clean: region(df_clean), axis=1)

df_full = pd.concat([df_clean, region], axis=1)

df_full.rename(columns={0: 'region', 'geography': 'city', 'Date': 'date'}, inplace=True)

df_full

# df_full.to_csv('./data_csv/updated_city_region.csv')
