==> 04-feature_engineering.ipynb <==
import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

raw = pd.read_csv('./data_csv/updated_city_region.csv', index_col=0)

raw

Dari feature selection, selected feature:
- year
- month
- type
- total_volume
- city

df = raw[['year', 'month', 'type', 'total_volume', 'city', 'average_price']]

df['total_volume'].describe()

bins = [84, 12170, 91688, 391257, 11324683]
labels = ['small', 'medium', 'large', 'wholesale']

df['category_vol'] = pd.cut(df['total_volume'], bins=bins, labels=labels)

df

df['category_vol'].value_counts()

## pengecekan jumlah baris setelah dibinning

6672+6672+6671+6670

df.drop(columns=['total_volume'], inplace=True)

df.head()

# df.to_csv('./data_csv/binned.csv')

# df.to_excel('./data_csv/selected.xlsx')
