==> 06-hyperparameter.ipynb <==
import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('./data_csv/binned.csv', index_col=0)

df.head()

### Splitting Data

X = df.drop(columns=['average_price'])
y = df['average_price']

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

X_train.columns

## 1. Hyperparameter LinearRegression dengan Polynomial preprocessing

from sklearn.linear_model import LinearRegression

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, PolynomialFeatures
from sklearn.compose import ColumnTransformer
import warnings
warnings.filterwarnings('ignore')

from sklearn.model_selection import GridSearchCV

transform = ColumnTransformer([
    ("encode", OneHotEncoder(), ['type', 'city', 'month', 'category_vol'])
], remainder = 'passthrough')

poly = Pipeline([
    ("transform", transform),
    ("poly", PolynomialFeatures())
])

lin_poly = Pipeline([
    ("prep_poly", poly),
    ("algo", LinearRegression())
])

lin_poly.get_params()

param_poly = {
    'prep_poly__poly__degree': [2, 3],
    "prep_poly__poly__include_bias" : ['True', 'False']
}

poly_gs = GridSearchCV(lin_poly, param_poly, cv=5, n_jobs=-1, verbose=1)

poly_gs.fit(X_train, y_train)

poly_gs.best_params_

preprocess = Pipeline([
    ("transform", transform),
    ("poly", PolynomialFeatures(degree=2, include_bias='True'))
])

model_poly = Pipeline([
    ("prep", preprocess),
    ("algo", LinearRegression())
])

model_poly.fit(X_train, y_train)

y_predPoly = model_poly.predict(X_test)

## evaluation

from sklearn.metrics import r2_score

r2_score(y_test, y_predPoly)

## 2. Hyperparameter DecisionTree

from sklearn.tree import DecisionTreeRegressor

### Preprocessing

- numerical features
    karena hanya 1 parameter numerik di variabel X - 'year', maka tidak perlu dilakukan preprocessing untuk variabel numerik
- categorical features: OneHotEncoder

X_train.info()

hype_DT = Pipeline([
    ("encode", transform),
    ("algo", DecisionTreeRegressor())
])

hype_DT.get_params()

param_DT = {
    'algo__max_depth': [None, 3, 5, 7],
    'algo__min_samples_leaf': range(1, 10, 2),
    'algo__min_samples_split': range(2, 20, 2),
    'algo__max_features': [None, 'log2', 'sqrt']
}

DT_GS = GridSearchCV(hype_DT, param_DT, cv=5, verbose=1)

DT_GS.fit(X_train, y_train)

DT_GS.best_params_

model_DT = Pipeline([
    ("encode", transform),
    ("algo", DecisionTreeRegressor(max_depth=None, max_features=None, min_samples_leaf= 1, min_samples_split= 2))
])

model_DT.fit(X_train, y_train)

y_predDT = model_DT.predict(X_test)

r2_score(y_test, y_predDT)

## 3. Hyperparameter dengan RandomForest

from sklearn.ensemble import RandomForestRegressorPolynomialFeatures

hype_RF = Pipeline([
    ("encode", transform),
    ("algo", RandomForestRegressor())
])

hype_RF.get_params()

params_RF = {
    'algo__n_estimators': range(100, 200, 50),
    'algo__max_depth': [None, 3, 5, 7],
    'algo__min_samples_leaf': range(1, 10, 2),
    'algo__min_samples_split': range(2, 20, 4),
    'algo__max_features': [None, 'sqrt']
}

RF_gs = GridSearchCV(hype_RF, params_RF, cv=5, n_jobs=-1, verbose=1)

RF_gs.fit(X_train, y_train)

RF_gs.best_params_

model_RF = Pipeline([
    ("encode", transform),
    ("algo", RandomForestRegressor(max_depth=None, max_features=None, min_samples_leaf= 1, min_samples_split= 2, n_estimators=150))
])

model_RF.fit(X_train, y_train)

y_predRF = model_RF.predict(X_test)

r2_score(y_test, y_predRF)

=========================================================================================

## evaluation

from sklearn.metrics import mean_absolute_error, mean_squared_error

summary = pd.DataFrame()

summary['model'] = ['Polynomial Regression', 'Decision Tree', 'Random Forest']

summary['MAE'] = [mean_absolute_error(y_test, y_predPoly), mean_absolute_error(y_test, y_predDT), mean_absolute_error(y_test, y_predRF)]
summary['MSE'] = [mean_squared_error(y_test, y_predPoly), mean_squared_error(y_test, y_predDT), mean_squared_error(y_test, y_predRF)]
summary['RMSE'] = [np.sqrt(mean_squared_error(y_test, y_predPoly)), np.sqrt(mean_squared_error(y_test, y_predDT)), np.sqrt(mean_squared_error(y_test, y_predRF))]
summary['r2_score'] = [r2_score(y_test, y_predPoly), r2_score(y_test, y_predDT), r2_score(y_test, y_predRF)]

summary

=========================================================================================

## Summary

Model dengan r2_score paling baik adalah Random Forest dengan parameter-parameter sbb:
- 'algo__max_depth': None,
- 'algo__max_features': None,
- 'algo__min_samples_leaf': 1,
- 'algo__min_samples_split': 2,
- 'algo__n_estimators': 150
