==> 08-model_predict.ipynb <==
import pickle
import pandas as pd

with open ('price_prediction', 'rb') as model:
    pred = pickle.load(model)

kolom = ['year', 'month', 'type', 'city', 'category_vol']

data = [[2020, 'September', 'organic', 'California', 'medium']]

# df = pd.DataFrame(data=data, columns=kolom)

predict = pred.predict(pd.DataFrame(data=data, columns=kolom)).round(3)

print(predict)
